requirements.txt
python-telegram-bot
python-dotenv
PyPDF2
regex
requests